Update the V1 model test suite
