--- v1/tests/models_tests/test_market.py
@@ -1,29 +1,0 @@
-from django.test import TestCase
-from market_nextdoor_api.models import Market
-
-class MarketModelTest(TestCase):
-  def setUp(self):
-    self.market_data = {
-        'market_name': 'Test Market',
-        'location': 'Test Location',
-        'details': 'Test Details',
-        'start_date': '2023-01-01',
-        'end_date': '2023-01-02',
-        'date_created': '2023-01-01T00:00:00Z',
-        'updated_at': '2023-01-02T00:00:00Z',
-    }
-
-    self.market = Market.objects.create(**self.market_data)
-
-def test_market_model(self):
-  market = Market.objects.get(id=self.market.id)
-  
-  self.assertEqual(market.market_name, self.market_data['market_name'])
-  self.assertEqual(market.location, self.market_data['location'])
-  self.assertEqual(market.details, self.market_data['details'])
-  self.assertEqual(str(market.start_date), self.market_data['start_date'])
-  self.assertEqual(str(market.end_date), self.market_data['end_date'])
-  self.assertEqual(market.date_created.isoformat(), self.market_data['date_created'])
-  self.assertEqual(market.updated_at.isoformat(), self.market_data['updated_at'])
-  self.assertEqual(str(market), self.market_data['market_name'])
-
