+++ v1/tests/models_tests/test_preorders.py
@@ -1,5 +1,5 @@
 from django.test import TestCase
-from market_nextdoor_api.models import Preorder, Customer, Item, Vendor
+from v1.models import Preorder, Customer, Item, Vendor
 
 class PreorderModelTest(TestCase):
   def setUp(self):
